requirements.txt
feedparser~=6.0.12
selenium~=4.41.0
webdriver-manager~=4.0.2
beautifulsoup4~=4.14.3
requests~=2.32.5
python-decouple~=3.8
dateparser~=1.3.0
playwright~=1.58.0
Django~=6.0.2
pyTelegramBotAPI~=4.31.0